Move warning from serialixation to validation
